requirements.txt
fastapi==0.115.12
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.3
PyQt5==5.15.10
requests==2.31.0
scikit-surprise==1.1.3
torch==2.2.0
pydantic==2.10.5
starlette==0.46.2
